// core-ui-salary-calculation-library/src/lib/data/repositories/usersManagement/users-api-urls.enum.ts
export enum UsersApiUrls{
    GetAreas = "/OrganizationUnit/GetAreasAndSubAreas",
    GetUserAssignedAreasAndSubAreas = "/Organizationunit/GetUserAssignedAreasAndSubAreas",
    GetUsers = "/Users/GetUsers",
    DisableUser = "/Users/DisableUser",
    EnableUser = "/Users/EnableUser",
    UpdateUser = "/Users/UpdateUser",
    GetRoles = "/Users/GetRoles",
    DeleteUser = "/Users/DeleteUser",
    GetAreaRoles = "/Users/GetAreaRoles"
}